chore: use format_gas and format_gas_throughput for gas logs
