+++ crates/exex/exex/Cargo.toml
@@ -19,6 +19,7 @@ reth-metrics.workspace = true
 reth-node-api.workspace = true
 reth-node-core.workspace = true
 reth-primitives.workspace = true
+reth-primitives-traits.workspace = true
 reth-provider.workspace = true
 reth-tasks.workspace = true
 reth-tracing.workspace = true
