+++ crates/stages/api/src/metrics/listener.rs
@@ -1,6 +1,6 @@
 use crate::{metrics::SyncMetrics, StageCheckpoint, StageId};
 use alloy_primitives::BlockNumber;
-use reth_primitives_traits::constants::MGAS_TO_GAS;
+use reth_primitives_traits::constants::MEGAGAS;
 use std::{
     future::Future,
     pin::Pin,
@@ -83,7 +83,7 @@ impl MetricsListener {
                 }
             }
             MetricEvent::ExecutionStageGas { gas } => {
-                self.sync_metrics.execution_stage.mgas_processed_total.increment(gas / MGAS_TO_GAS)
+                self.sync_metrics.execution_stage.mgas_processed_total.increment(gas / MEGAGAS)
             }
         }
     }
