+++ crates/stages/stages/Cargo.toml
@@ -24,6 +24,7 @@ reth-evm.workspace = true
 reth-exex.workspace = true
 reth-network-p2p.workspace = true
 reth-primitives.workspace = true
+reth-primitives-traits.workspace = true
 reth-provider.workspace = true
 reth-execution-types.workspace = true
 reth-prune-types.workspace = true
